Add files via upload
Rice Up is a project focused on improving rice crop management in Egypt. By implementing an IoT-based system for early detection and control of rice diseases, we aim to minimize crop losses and increase rice production. Our user-friendly mobile application provides farmers with real-time data and insights about the health of their rice crops. With our innovative approach, we strive to contribute to Egypt's self-sufficiency in rice production and enhance the overall sustainability of the rice industry.